+++ crates/tron/src/lib.rs
@@ -20,7 +20,7 @@ pub mod provider;
 pub mod tvm;
 mod wallet;
 
-pub use alloy_primitives::{Bytes, Log, U256};
+pub use alloy_primitives::{Bytes, Log, B256, U256};
 pub use asset::TronAsset;
 pub use chain::TronChain;
 pub use chains::TronChainInfo;